Shut up most of the overly pendantic warnings in VS.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@16587 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/class_modules/sc_death_knight.cpp
@@ -362,7 +362,7 @@ public:
       slot[2].paired_rune = &slot[ 3 ]; slot[ 3 ].paired_rune = &slot[ 2 ];
       slot[4].paired_rune = &slot[ 5 ]; slot[ 5 ].paired_rune = &slot[ 4 ];
       // give each rune a slot number
-      for ( size_t i = 0; i < slot.size(); ++i ) { slot[ i ].slot_number = i; }
+      for ( size_t i = 0; i < slot.size(); ++i ) { slot[ i ].slot_number = static_cast<int>( i ); }
     }
     void reset() { for ( size_t i = 0; i < slot.size(); ++i ) slot[ i ].reset(); }
   } _runes;
